--- src/middlewares/delete-url.js
@@ -1,5 +1,0 @@
-import UrlShorten from '../models/UrlShorten';
-
-export const deleteUrl = (req, res) => {
-  return;
-}
