+++ src/middlewares/middlewares.js
@@ -1,5 +1,4 @@
 import { renderLandingPage } from './renderLandingPage';
-import { trimUrl } from './trim-url';
-import { deleteUrl } from './delete-url';
+import { checkUrl } from './urlMiddleware';
 
-export { renderLandingPage, trimUrl, deleteUrl };
+export { renderLandingPage, checkUrl };
