+++ src/middlewares/renderLandingPage.js
@@ -1,5 +1,12 @@
 import UrlShorten from '../models/UrlShorten';
 
+
+/**
+ * This function renders the landing page and gets list of user trimmed urls
+ * @param {object} req
+ * @param {object} res
+ * @returns {object} response object with trimmed url
+ */
 export const renderLandingPage = (req, res) => {
 	UrlShorten.find({
     created_by: req.cookies.userID //Find all clips created by this user.
