--- src/middlewares/trim-url.js
@@ -1,5 +1,0 @@
-import UrlShorten from '../models/UrlShorten';
-
-export const trimUrl = (req, res) => {
-  return;
-}
